requirements.txt
Django==2.2
djangorestframework==3.9.2
django-filter==2.1.0
requests==2.21.0
python-dateutil==2.8.0
